# add a layer of caching for valid ZKPs and their statements