--- a/aptos-move/aptos-vm/src/lib.rs
+++ b/aptos-move/aptos-vm/src/lib.rs
@@ -109,7 +109,10 @@ pub mod aptos_vm;
 pub mod block_executor;
 mod errors;
 pub mod gas;
+#[cfg(not(feature = "testing"))]
 mod keyless_validation;
+#[cfg(feature = "testing")]
+pub mod keyless_validation;
 pub mod move_vm_ext;
 pub mod natives;
 pub mod sharded_block_executor;