**Note:** Closing this and breaking it down into smaller PRs.

## TODOs

 - [ ] add `InstalledEPK` resource in `keyless_account.move`
 - [ ] avoid expensive VK deserialization
 - <strike>invalidate ZKP cache after VK change</strike>
    * If we install the EPKs, we will not need a ZKP cache anymore.
    * Clearing the installed EPKs after the VK change achieves nothing in terms of security.
 - [ ] faster Poseidon hashing

## Description

This PR adds a cache of successfully-verified ZKPs, some tests & benchmarks.

### 1. ZKP cache (w/ VK change detection)

Add a cache of previously-verified ZKPs and their corresponding statements. This speeds up keyless TXN verification, ensuring that a valid ZKP does not get reverified unnecessarily.

The tricky part is that this cache must be invalidated when the VK changes.

### 2. Authenticator size tests

Also, tested the byte sizes of a serialized keyless pubkey and signature:

```
OpenID sizes
------------
KeylessSignature BCS size: 1033
KeylessPublicKey BCS size: 61

(Signature size would have been 294 if the extra_field was set to None.)

Groth16 sizes
------------
KeylessSignature BCS size: 318
KeylessPublicKey BCS size: 61
```

### 3. ZKP benchmarks

```
VK deserialization time: 1.92ms
 + Without validation: 1.00ms
Public inputs hash time: 1.609917ms
Proof deserialization time: 418.5µs
Proof verification time: 1.069708ms
------------------------------------
Total time: 5.01ms
```

### 4. Install EPK after successful verification

## Type of Change
- [ ] New feature
- [ ] Bug fix
- [ ] Breaking change
- [x] Performance improvement
- [x] Refactoring
- [x] Tests
- [x] Benchmarks
- [ ] Dependency update
- [ ] Documentation update

## Which Components or Systems Does This Change Impact?
- [x] Validator Node
- [x] Full Node (API, Indexer, etc.)
- [x] Move/Aptos Virtual Machine
- [ ] Aptos Framework
- [ ] Aptos CLI/SDK
- [ ] Developer Infrastructure
- [ ] Other (specify)

## How Has This Been Tested?

First, the caching logic is rather simple:

```rust
// If the ZKP has not successfully verified over this statment in the past,
// manually verify it & cache it. Otherwise, skip.
let entry = (groth16_and_stmt, vk_hash);
if matches!(ZKP_CACHE.get(&entry), None | Some(false)) {
    let result = zksig.verify_groth16_proof(public_inputs_hash, &pvk);

    if result.is_ok() {
        ZKP_CACHE.insert(entry, true);
    }

    result
        .map_err(|_| invalid_signature!("Proof verification failed"))?;
}
```

Nonetheless, we will test a couple of things:

1. Previously **un**seen ZKP does **not** hit the cache.
2. Previously seen ZKP hits the cache.
3. VK rotation clears the cache.

## Key Areas to Review

 > Identify any critical parts of the code that require special attention or understanding. Explain why these parts are crucial to the functionality or architecture of the project.

1. The use of a `Lazy<Cache>` should be checked. I am not sure if it should have been a `Lazy<Arc<Cache>>`
4. The changes in `keyless_validation.rs` should be checked.
5. The use of the `quick_cache` crate, as recommended by @gelash as an initial good-enough cache implementation.
6. The use of a global static for the ZKP cache in Aptos VM, as opposed to a per-VM cache (cc @gelash)

 > Point out any areas where complex logic has been implemented. Provide a brief explanation of the logic and your approach to make it easier for reviewers to follow.

No complex logic: just check if the proof & the statement were previously-verified. If not, verify them. Otherwise, skip expensive verification.

 > Highlight any areas where you are particularly concerned or unsure about the code's impact on the change. This can include potential performance or security issues, or compatibility with existing features.

Particularly worried about:
 1. The global static for the ZKP cache in Aptos VM, as opposed to a per-VM cache. We do want the cache to be longer-lived than an `AptosVM`, since later blocks might include the same ZKP and we'd want the cache to hit.

(cc @gelash)

## Checklist
- [x] I have read and followed the [CONTRIBUTING](https://github.com/aptos-labs/aptos-core/blob/main/CONTRIBUTING.md) doc
- [ ] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I identified and added all stakeholders and component owners affected by this change as reviewers
- [ ] I tested both happy and unhappy path of the functionality
- [x] I have made corresponding changes to the documentation

## Dismissed approaches

### Alternative ways of invalidating the ZKP cache after a VK change

We need a way to make sure the ZKP cache is invalidated after a VK change. Otherwise, old, potentially-insecure ZKPs will continue verifying after the VK change.

A potentially-working approach is in 79921ada9d018bf1476534877dc5af41c8d34ea0:

```rust
let vk = get_groth16_vk_onchain(resolver)?;

// If the VK has changed, the cache MUST be invalidated
let mut last_seen_vk = LAST_SEEN_VK.lock().unwrap();
if *vk != *last_seen_vk {
    ZKP_CACHE.clear();
    *last_seen_vk = vk.clone();
}
```

However, there are some general concerns from @zekun000 who wants to avoid complexity.

Also, the use of a `Lazy<Mutex<Groth16VerificationKey>>` mutex global to see if the VK has changed might be problematic from a performance standpoint since it seems to require locking the VK during keyless validation, which effectively serializes the verification of all keyless TXNs.

Might it be better to store the VK in the `AptosVM` somewhere? Is there already a standard way to detect changes to on-chain state & update `AptosVM` state?

An alternative proposed by @zekun000 is to trigger a reconfiguration after a VK change, in which case there are proper mechanisms to refresh the `AptosVM`'s view of the VK.

Another suggestion from @zekun000 was to include the VK (hash) in the cache next to the successfully-verified ZKP. But this increases the size of the cache.

<!-- Thank you for your contribution! -->